Internal images
Fixed the image references so we're not calling them from an outside source any more. Turns out we can import them to different components, which keeps things nice and tidy.

Have added a handful of fair use ads from Wikimedia Commons to test this though they may or may not make the cut.